refactor: extract separator planning
